+++ tests/testthat/test-standardizedSolution_boot_ci_print.R
@@ -1,6 +1,6 @@
-# skip_on_cran()
-# skip_if(!interactive(),
-#         message = "standardizedSolution_boot_ci not tested if not interactive")
+skip_on_cran()
+skip_if(!interactive(),
+        message = "print method not tested if not interactive")
 
 library(testthat)
 library(semhelpinghands)
@@ -13,7 +13,10 @@ n <- 1000
 X <- runif(n) - .5
 M <- 0.20*X + rnorm(n)
 Y <- 0.17*M + rnorm(n)
-Data <- data.frame(X = X, Y = Y, M = M)
+gp <- sample(c("Group1", "Group2"),
+             n,
+             replace = TRUE)
+Data <- data.frame(X = X, Y = Y, M = M, gp = gp)
 model <- ' # direct effect
              Y ~ c*X
            # mediator
@@ -24,14 +27,42 @@ model <- ' # direct effect
            # total effect
              total := c + (a*b)
          '
+model2 <- ' # direct effect
+             Y ~ c(c1, c2)*X
+           # mediator
+             M ~ c(a1, a2)*X
+             Y ~ c(b1, b2)*M
+           # indirect effect (a*b)
+             a1b1 := a1*b2
+             a2b2 := a2*b2
+           # total effect
+             total1 := c1 + (a1*b1)
+             total2 := c2 + (a2*b2)
+         '
 set.seed(1234)
-# One bootstrap replication failed. Kept intentionally.
 suppressWarnings(system.time(fit <- sem(model,
                        data = Data,
                        se = "boot",
-                       bootstrap = 100)))
+                       bootstrap = 50)))
+set.seed(1234)
+suppressWarnings(system.time(fit2 <- sem(model2,
+                       data = Data,
+                       se = "boot",
+                       bootstrap = 50,
+                       group = "gp")))
 
 ci_boot <- standardizedSolution_boot_ci(fit, save_boot_est_std = TRUE)
+ci_boot2 <- standardizedSolution_boot_ci(fit2, save_boot_est_std = TRUE)
+
+print(ci_boot, nd = 5)
+print(ci_boot, output = "text")
+print(ci_boot, output = "text", standardized_only = FALSE)
+
+print(ci_boot2, nd = 5)
+print(ci_boot2, output = "text")
+print(ci_boot2, output = "text", standardized_only = FALSE)
+
+
 
 get_std <- function(object) {
     lavaan::standardizedSolution(object)$est.std
@@ -59,16 +90,4 @@ test_that("Compare boot estimates directly", {
         boot_ci_test,
         ignore_attr = TRUE
       )
-  })
-
-# Test store_boot_est_std()
-
-tmp <- store_boot_est_std(fit)
-tmp_boot_est_std <- get_boot_est_std(tmp)
-test_that("store_boot_est_std", {
-    expect_equal(tmp_boot_est_std,
-                 attr(ci_boot, "boot_est_std"),
-                 ignore_attr = TRUE)
-    expect_equal(tmp@external$shh_boot_est_std_type,
-                 "std.all")
   })